Trait has been converted
